Updates to add Miles, Lbs, and Meditation sensors
This is a clunky approach, just adding sensors instead of linking it to the original settings of HA, but it works. I had forked the repository at https://github.com/vmanuel/hacs-google-fit/issues which was based off of yours and the Miles and Lbs sensors I added as a pull request there, but figured I should add one here as well. Feel free to reject if that's not the direction you're looking to go.

Either way, thanks for a great start on this; really was easy to extend!